requirements.txt
aiofiles==0.7.0
anyio==3.3.0
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
datefinder==0.7.1
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
fastapi==0.68.1
greenlet==1.1.1
gunicorn==20.1.0
h11==0.12.0
idna==3.2
IMAPClient==2.2.0
ipykernel==6.2.0
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.2
kiwisolver==1.3.2
mail-parser==3.15.0
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parse==1.19.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pyTelegramBotAPI==4.0.0
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.2.1
regex==2021.8.28
requests==2.26.0
requests-unixsocket==0.2.0
Send2Trash==1.8.0
simplejson==3.17.5
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
starlette==0.14.2
terminado==0.11.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
